[Update] Refactor Class name and package to suitable locations
